requirements.txt
anthropic
google-genai
streamlit
python-dotenv
Pillow
Jinja2
playwright
pandas
